+++ .replit
@@ -1,0 +1,1 @@
+run = ['npm install --only=prod, 'npm start']
