+++ index.js
@@ -1,9 +1,14 @@
 const express = require('express');
-var cors = require('cors');
+const cors = require('cors');
 const axios = require('axios');
-const cheerio = require('cheerio');
 const dotenv = require('dotenv');
 const fs = require('fs');
+
+dotenv.config();
+
+const { LOG_FILE, OMDB_URL, contentAdvisoryURL } = require('./constants');
+
+const { getContentAdvisory, getTitlesFromIMDB } = require('./helper-functions');
 
 const noop = () => {};
 
@@ -13,140 +18,10 @@ const noop = () => {};
 const app = express();
 app.use(express.json());
 app.use(cors());
-
-dotenv.config();
+const PORT = process.env.PORT || 3000;
+const IS_DEV = process.env.IS_DEV;
 
 // API URL : https://imdb-parental-advisory.xsaudahmed.repl.co
-
-/**
- * CONSTANTS
- */
-const SELECTORS = {
-  TITLES_SECTION: '[name="tt"]',
-  FOUND_RESULT_ROW: '.findResult',
-};
-
-const LOG_FILE = 'requests.log';
-const API_KEY = process.env.OMDB_API_KEY;
-const OMDB_URL = `http://www.omdbapi.com/?apikey=${API_KEY}`;
-
-/**
- * URLS
- */
-const searchURL = (titleName) =>
-  `https://www.imdb.com/find?q=${encodeURIComponent(
-    titleName
-  )}&s=tt&ref_=fn_al_tt_mr`;
-const contentAdvisoryURL = (titleID) =>
-  `https://www.imdb.com/title/${titleID}/parentalguide?ref_=tt_stry_pg`;
-
-/**
- * HELPER FUNCTIONS
- */
-function getSectionInformation(isSpoilersSection) {
-  return function (section) {
-    try {
-      const result = {};
-
-      const currentSection = cheerio.load(section);
-
-      const sectionName = currentSection('h4').text();
-
-      if (!isSpoilersSection) {
-        const [advisory, ...entries] = Array.from(currentSection('ul li'));
-
-        const advisorySection = cheerio.load(advisory);
-        const summary = advisorySection(
-          '.advisory-severity-vote__container span'
-        ).html();
-        const voteCount = advisorySection(
-          '.advisory-severity-vote__container a'
-        )
-          .text()
-          .trim();
-
-        result.advisory = {
-          summary,
-          voteCount,
-        };
-        result.entries = entries.map((entry) => entry.children[0].data.trim());
-      } else {
-        result.entries = Array.from(currentSection('ul li')).map((entry) =>
-          entry.children[0].data.trim()
-        );
-      }
-
-      result.sectionName = sectionName;
-
-      return result;
-    } catch (err) {
-      return { entries: [] };
-    }
-  };
-}
-
-async function getContentAdvisory(titleID) {
-  const res = await axios.get(contentAdvisoryURL(titleID));
-  const wholeDocument = cheerio.load(res.data);
-
-  const parentalGuideSection = wholeDocument('[id^="advisory"]').slice(0, 5);
-  const spoilersGuideSection = wholeDocument('#advisory-spoilers section');
-  const title = wholeDocument('h3[itemprop="name"]').text();
-  console.log(title);
-  const parentalGuide = Array.from(parentalGuideSection).map(
-    getSectionInformation(false)
-  );
-
-  const spoilersGuide = Array.from(spoilersGuideSection).map(
-    getSectionInformation(true)
-  );
-
-  return {
-    title,
-    parentalGuide,
-    spoilersGuide,
-  };
-}
-
-async function getTitlesFromIMDB(titleName) {
-  try {
-    const res = await axios.get(searchURL(titleName));
-
-    const wholeDocument = cheerio.load(res.data);
-    const scrapedTitles = wholeDocument(SELECTORS.TITLES_SECTION)
-      .parent()
-      .parent()
-      .html();
-
-    if (scrapedTitles) {
-      const titleSection = cheerio.load(scrapedTitles);
-
-      const resultRows = titleSection(SELECTORS.FOUND_RESULT_ROW)
-        .toArray()
-        .slice(0, 15);
-
-      const responseArray = resultRows.map((row) => {
-        const rowChildren = cheerio.load(row.children);
-
-        const id = rowChildren('a')[0].attribs.href.split('/')[2];
-        const imageURL = rowChildren('img')[0].attribs.src;
-        const title = cheerio.load(row).text();
-
-        return {
-          id,
-          imageURL,
-          title,
-        };
-      });
-
-      return responseArray;
-    } else {
-      return [];
-    }
-  } catch (error) {
-    console.log(error);
-  }
-}
 
 /**
  * ROUTES
@@ -193,7 +68,7 @@ app.post('/getRatings', async (req, res) => {
 
   try {
     if (titleId) {
-      const response = await axios.post(OMDB_URL + `&i=${titleId}`);
+      const response = await axios.post(`${OMDB_URL}&i=${titleId}`);
 
       res.send(response.data);
     } else {
@@ -207,6 +82,8 @@ app.post('/getRatings', async (req, res) => {
   }
 });
 
-app.listen(3000, () => {
-  console.log('server started');
+app.listen(PORT, () => {
+  if (IS_DEV) {
+    console.log(`server started on http://localhost:${PORT}`);
+  }
 });
