+++ start.sh
@@ -1,0 +1,2 @@
+npm install --only=prod
+npm start
